Move createEnvOption to utils/cpUtils.ts
